Migrate from gevent-psycopg to psycogreen
pip now requires that packages be downloaded over HTTPS. Installing
gevent-psycopg triggers a download over HTTP. The recommended solution to the
HTTP problem is to upgrade to the newest version of the offending package.
gevent-psycopg is deprecated in favor or psycogreen, so I'm migrating to that
instead.

I'm not sure what we're using this for :(